+++ lib/src/internal/debug.dart
@@ -1,0 +1,6 @@
+import 'package:meta/meta.dart';
+
+@internal
+class DockingDebug {
+  static const bool dropAreaVisible = false;
+}