+++ lib/src/internal/widgets/docking_tabs_widget.dart
@@ -1,9 +1,11 @@
 import 'dart:math' as math;
-import 'package:docking/src/docking_drag.dart';
+
 import 'package:docking/src/docking_buttons_builder.dart';
+import 'package:docking/src/docking_drag.dart';
 import 'package:docking/src/docking_icons.dart';
-import 'package:docking/src/internal/widgets/draggable_widget.dart';
-import 'package:docking/src/internal/widgets/drop_widget.dart';
+import 'package:docking/src/internal/widgets/draggable_config_mixin.dart';
+import 'package:docking/src/internal/widgets/drop/content_wrapper.dart';
+import 'package:docking/src/internal/widgets/drop/drop_feedback_widget.dart';
 import 'package:docking/src/layout/docking_layout.dart';
 import 'package:docking/src/on_item_close.dart';
 import 'package:docking/src/on_item_selection.dart';
@@ -11,11 +13,11 @@ import 'package:flutter/material.dart';
 import 'package:tabbed_view/tabbed_view.dart';
 
 /// Represents a widget for [DockingTabs].
-class DockingTabsWidget extends DraggableWidget {
+class DockingTabsWidget extends StatefulWidget {
   DockingTabsWidget(
       {Key? key,
       required this.layout,
-      required DockingDrag dockingDrag,
+      required this.dockingDrag,
       required this.dockingTabs,
       this.onItemSelection,
       this.onItemClose,
@@ -23,7 +25,7 @@ class DockingTabsWidget extends DraggableWidget {
       this.dockingButtonsBuilder,
       required this.maximizableTab,
       required this.maximizableTabsArea})
-      : super(key: key, dockingDrag: dockingDrag);
+      : super(key: key);
 
   final DockingLayout layout;
   final DockingTabs dockingTabs;
@@ -33,11 +35,20 @@ class DockingTabsWidget extends DraggableWidget {
   final DockingButtonsBuilder? dockingButtonsBuilder;
   final bool maximizableTab;
   final bool maximizableTabsArea;
+  final DockingDrag dockingDrag;
+
+  @override
+  State<StatefulWidget> createState() => DockingTabsWidgetState();
+}
+
+class DockingTabsWidgetState extends State<DockingTabsWidget>
+    with DraggableConfigMixin {
+  DropPosition? _activeDropPosition;
 
   @override
   Widget build(BuildContext context) {
     List<TabData> tabs = [];
-    dockingTabs.forEach((child) {
+    widget.dockingTabs.forEach((child) {
       Widget content = child.widget;
       if (child.globalKey != null) {
         content = KeyedSubtree(child: content, key: child.globalKey);
@@ -47,20 +58,22 @@ class DockingTabsWidget extends DraggableWidget {
         buttons = [];
         buttons.addAll(child.buttons!);
       }
-      final bool maximizable =
-          child.maximizable != null ? child.maximizable! : this.maximizableTab;
+      final bool maximizable = child.maximizable != null
+          ? child.maximizable!
+          : widget.maximizableTab;
       if (maximizable) {
         if (buttons == null) {
           buttons = [];
         }
-        if (layout.maximizedArea != null && layout.maximizedArea == child) {
+        if (widget.layout.maximizedArea != null &&
+            widget.layout.maximizedArea == child) {
           buttons.add(TabButton(
               icon: IconProvider.path(DockingIcons.restore),
-              onPressed: () => layout.restore()));
+              onPressed: () => widget.layout.restore()));
         } else {
           buttons.add(TabButton(
               icon: IconProvider.data(Icons.web_asset_sharp),
-              onPressed: () => layout.maximizeDockingItem(child)));
+              onPressed: () => widget.layout.maximizeDockingItem(child)));
         }
       }
       tabs.add(TabData(
@@ -74,64 +87,94 @@ class DockingTabsWidget extends DraggableWidget {
     });
     TabbedViewController controller = TabbedViewController(tabs);
     controller.selectedIndex =
-        math.min(dockingTabs.selectedIndex, tabs.length - 1);
+        math.min(widget.dockingTabs.selectedIndex, tabs.length - 1);
 
     Widget tabbedView = TabbedView(
         controller: controller,
         tabsAreaButtonsBuilder: _tabsAreaButtonsBuilder,
         onTabSelection: (int? index) {
           if (index != null) {
-            dockingTabs.selectedIndex = index;
-            if (onItemSelection != null) {
-              onItemSelection!(dockingTabs.childAt(index));
+            widget.dockingTabs.selectedIndex = index;
+            if (widget.onItemSelection != null) {
+              widget.onItemSelection!(widget.dockingTabs.childAt(index));
             }
           }
         },
         tabCloseInterceptor: _tabCloseInterceptor,
-        draggableTabBuilder: (int tabIndex, TabData tab, Widget tabWidget) {
-          return buildDraggable(dockingTabs.childAt(tabIndex), tabWidget);
+        onDraggableBuild:
+            (TabbedViewController controller, int tabIndex, TabData tabData) {
+          return buildDraggableConfig(
+              dockingDrag: widget.dockingDrag, tabData: tabData);
         },
-        onTabClose: _onTabClose);
-    if (dockingDrag.enable) {
-      return DropWidget.tabs(layout, dockingTabs, tabbedView);
+        onTabClose: _onTabClose,
+        contentBuilder: (context, tabIndex) => TabsContentWrapper(
+            listener: _updateActiveDropPosition,
+            layout: widget.layout,
+            dockingTabs: widget.dockingTabs,
+            child: controller.tabs[tabIndex].content!),
+        onBeforeDropAccept: _onBeforeDropAccept);
+    if (widget.dockingDrag.enable) {
+      return DropFeedbackWidget(
+          dropPosition: _activeDropPosition, child: tabbedView);
     }
     return tabbedView;
   }
 
+  void _updateActiveDropPosition(DropPosition? dropPosition) {
+    if (_activeDropPosition != dropPosition) {
+      setState(() {
+        _activeDropPosition = dropPosition;
+      });
+    }
+  }
+
+  bool _onBeforeDropAccept(
+      DraggableData source, TabbedViewController target, int newIndex) {
+    DockingItem dockingItem = source.tabData.value;
+    widget.layout.moveItem(
+        draggedItem: dockingItem,
+        targetArea: widget.dockingTabs,
+        dropIndex: newIndex);
+    return true;
+  }
+
   List<TabButton> _tabsAreaButtonsBuilder(BuildContext context, int tabsCount) {
     List<TabButton> buttons = [];
-    if (dockingButtonsBuilder != null) {
-      buttons.addAll(dockingButtonsBuilder!(context, dockingTabs, null));
+    if (widget.dockingButtonsBuilder != null) {
+      buttons.addAll(
+          widget.dockingButtonsBuilder!(context, widget.dockingTabs, null));
     }
-    final bool maximizable = dockingTabs.maximizable != null
-        ? dockingTabs.maximizable!
-        : this.maximizableTabsArea;
+    final bool maximizable = widget.dockingTabs.maximizable != null
+        ? widget.dockingTabs.maximizable!
+        : widget.maximizableTabsArea;
     if (maximizable) {
-      if (layout.maximizedArea != null && layout.maximizedArea == dockingTabs) {
+      if (widget.layout.maximizedArea != null &&
+          widget.layout.maximizedArea == widget.dockingTabs) {
         buttons.add(TabButton(
             icon: IconProvider.path(DockingIcons.restore),
-            onPressed: () => layout.restore()));
+            onPressed: () => widget.layout.restore()));
       } else {
         buttons.add(TabButton(
             icon: IconProvider.data(Icons.web_asset_sharp),
-            onPressed: () => layout.maximizeDockingTabs(dockingTabs)));
+            onPressed: () =>
+                widget.layout.maximizeDockingTabs(widget.dockingTabs)));
       }
     }
     return buttons;
   }
 
   bool _tabCloseInterceptor(int tabIndex) {
-    if (itemCloseInterceptor != null) {
-      return itemCloseInterceptor!(dockingTabs.childAt(tabIndex));
+    if (widget.itemCloseInterceptor != null) {
+      return widget.itemCloseInterceptor!(widget.dockingTabs.childAt(tabIndex));
     }
     return true;
   }
 
   void _onTabClose(int tabIndex, TabData tabData) {
-    DockingItem dockingItem = dockingTabs.childAt(tabIndex);
-    layout.removeItem(item: dockingItem);
-    if (onItemClose != null) {
-      onItemClose!(dockingItem);
+    DockingItem dockingItem = widget.dockingTabs.childAt(tabIndex);
+    widget.layout.removeItem(item: dockingItem);
+    if (widget.onItemClose != null) {
+      widget.onItemClose!(dockingItem);
     }
   }
 }