+++ lib/src/layout/docking_layout.dart
@@ -168,6 +168,11 @@ abstract class DockingParentArea extends DockingArea {
   /// Gets a child for a given index.
   DockingArea childAt(int index) => _children[index];
 
+  /// The first index of [dockingArea] in this container.
+  ///
+  /// Returns -1 if [dockingArea] is not found.
+  int indexOf(DockingArea dockingArea) => _children.indexOf(dockingArea);
+
   /// Whether the [DockingParentArea] contains a child equal to [area].
   bool contains(DockingArea area) {
     return _children.contains(area);
@@ -432,7 +437,7 @@ enum DockingAreaType { item, tabs, row, column }
 
 /// Represents all positions available for a drop event that will
 /// rearrange the layout.
-enum DropPosition { top, bottom, left, right, center }
+enum DropPosition { top, bottom, left, right }
 
 /// Represents a layout.
 ///
@@ -581,13 +586,15 @@ class DockingLayout extends ChangeNotifier {
   void moveItem(
       {required DockingItem draggedItem,
       required DropArea targetArea,
-      required DropPosition dropPosition}) {
+      DropPosition? dropPosition,
+      int? dropIndex}) {
     //TODO maximize test
     _rebuild([
       MoveItem(
           draggedItem: draggedItem,
           targetArea: targetArea,
-          dropPosition: dropPosition)
+          dropPosition: dropPosition,
+          dropIndex: dropIndex)
     ]);
   }
 
@@ -609,17 +616,23 @@ class DockingLayout extends ChangeNotifier {
   void addItemOn(
       {required DockingItem newItem,
       required DropArea targetArea,
-      required DropPosition dropPosition}) {
+      DropPosition? dropPosition,
+      int? dropIndex}) {
     //TODO maximize test
     _rebuild([
       AddItem(
-          newItem: newItem, targetArea: targetArea, dropPosition: dropPosition)
+          newItem: newItem,
+          targetArea: targetArea,
+          dropPosition: dropPosition,
+          dropIndex: dropIndex)
     ]);
   }
 
   /// Adds a DockingItem to the root of this layout.
   void addItemOnRoot(
-      {required DockingItem newItem, required DropPosition dropPosition}) {
+      {required DockingItem newItem,
+      DropPosition? dropPosition,
+      int? dropIndex}) {
     if (root == null) {
       throw StateError('Root is null');
     }
@@ -629,7 +642,8 @@ class DockingLayout extends ChangeNotifier {
         AddItem(
             newItem: newItem,
             targetArea: targetArea,
-            dropPosition: dropPosition)
+            dropPosition: dropPosition,
+            dropIndex: dropIndex)
       ]);
     } else {
       throw StateError('Root is not a DropArea');
@@ -690,9 +704,4 @@ class DockingLayout extends ChangeNotifier {
       parentArea.forEach((child) => _fetchAreas(areas, child));
     }
   }
-
-  void deserialize(
-      String layout, WidgetDeserialization widgetDeserialization) {}
 }
-
-typedef WidgetDeserialization = Widget Function(dynamic itemId);