+++ test/move_test.dart
@@ -9,11 +9,11 @@ void main() {
     test('draggedItem == targetArea ', () {
       DockingItem item = dockingItem('a');
       DockingLayout layout = DockingLayout(root: item);
-      expect(() => moveItem(layout, item, item, DropPosition.bottom),
+      expect(() => moveItemToPosition(layout, item, item, DropPosition.bottom),
           sameDraggedItemAndTargetAreaException());
     });
 
-    test('nested tabbed panel', () {
+    test('nested tabbed panel - 0', () {
       DockingItem itemA = dockingItem('a');
       DockingItem itemB = dockingItem('b');
       DockingItem itemC = dockingItem('c');
@@ -21,8 +21,20 @@ void main() {
       DockingRow row = DockingRow([itemA, tabs]);
       DockingLayout layout = DockingLayout(root: row);
 
-      expect(() => moveItem(layout, itemA, itemB, DropPosition.center),
-          throwsArgumentError);
+      expect(
+          () => moveItemToIndex(layout, itemA, itemB, 0), throwsArgumentError);
+    });
+
+    test('nested tabbed panel - 1', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingItem itemC = dockingItem('c');
+      DockingTabs tabs = DockingTabs([itemB, itemC]);
+      DockingRow row = DockingRow([itemA, tabs]);
+      DockingLayout layout = DockingLayout(root: row);
+
+      expect(
+          () => moveItemToIndex(layout, itemA, itemB, 1), throwsArgumentError);
     });
   });
 
@@ -33,7 +45,7 @@ void main() {
       DockingRow row = DockingRow([itemA, itemB]);
       DockingLayout layout = DockingLayout(root: row);
 
-      moveItem(layout, itemA, itemB, DropPosition.left);
+      moveItemToPosition(layout, itemA, itemB, DropPosition.left);
 
       testHierarchy(layout, 'R(Ia,Ib)');
     });
@@ -44,7 +56,7 @@ void main() {
       DockingRow row = DockingRow([itemA, itemB]);
       DockingLayout layout = DockingLayout(root: row);
 
-      moveItem(layout, itemA, itemB, DropPosition.right);
+      moveItemToPosition(layout, itemA, itemB, DropPosition.right);
 
       testHierarchy(layout, 'R(Ib,Ia)');
     });
@@ -55,7 +67,7 @@ void main() {
       DockingRow row = DockingRow([itemA, itemB]);
       DockingLayout layout = DockingLayout(root: row);
 
-      moveItem(layout, itemA, itemB, DropPosition.top);
+      moveItemToPosition(layout, itemA, itemB, DropPosition.top);
 
       testHierarchy(layout, 'C(Ia,Ib)');
     });
@@ -66,7 +78,7 @@ void main() {
       DockingRow row = DockingRow([itemA, itemB]);
       DockingLayout layout = DockingLayout(root: row);
 
-      moveItem(layout, itemA, itemB, DropPosition.bottom);
+      moveItemToPosition(layout, itemA, itemB, DropPosition.bottom);
 
       testHierarchy(layout, 'C(Ib,Ia)');
     });
@@ -77,12 +89,23 @@ void main() {
       DockingRow row = DockingRow([itemA, itemB]);
       DockingLayout layout = DockingLayout(root: row);
 
-      moveItem(layout, itemA, itemB, DropPosition.center);
+      moveItemToIndex(layout, itemA, itemB, 0);
+
+      testHierarchy(layout, 'T(Ia,Ib)');
+    });
+
+    test('row - to tabs 2', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingRow row = DockingRow([itemA, itemB]);
+      DockingLayout layout = DockingLayout(root: row);
+
+      moveItemToIndex(layout, itemA, itemB, 1);
 
       testHierarchy(layout, 'T(Ib,Ia)');
     });
 
-    test('row - to tabs 2', () {
+    test('row - to tabs 3', () {
       DockingItem itemA = dockingItem('a');
       DockingItem itemB = dockingItem('b');
       DockingItem itemC = dockingItem('c');
@@ -90,12 +113,38 @@ void main() {
       DockingRow row = DockingRow([itemA, tabs]);
       DockingLayout layout = DockingLayout(root: row);
 
-      moveItem(layout, itemA, tabs, DropPosition.center);
+      moveItemToIndex(layout, itemA, tabs, 0);
+
+      testHierarchy(layout, 'T(Ia,Ib,Ic)');
+    });
+
+    test('row - to tabs 4', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingItem itemC = dockingItem('c');
+      DockingTabs tabs = DockingTabs([itemB, itemC]);
+      DockingRow row = DockingRow([itemA, tabs]);
+      DockingLayout layout = DockingLayout(root: row);
+
+      moveItemToIndex(layout, itemA, tabs, 1);
+
+      testHierarchy(layout, 'T(Ib,Ia,Ic)');
+    });
+
+    test('row - to tabs 5', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingItem itemC = dockingItem('c');
+      DockingTabs tabs = DockingTabs([itemB, itemC]);
+      DockingRow row = DockingRow([itemA, tabs]);
+      DockingLayout layout = DockingLayout(root: row);
+
+      moveItemToIndex(layout, itemA, tabs, 2);
 
       testHierarchy(layout, 'T(Ib,Ic,Ia)');
     });
 
-    test('complex 1', () {
+    test('complex 1 a', () {
       DockingItem itemA = dockingItem('a');
       DockingItem itemB = dockingItem('b');
       DockingItem itemC = dockingItem('c');
@@ -107,12 +156,29 @@ void main() {
       DockingColumn column = DockingColumn([row, tabs]);
       DockingLayout layout = DockingLayout(root: column);
 
-      moveItem(layout, itemA, itemC, DropPosition.center);
+      moveItemToIndex(layout, itemA, itemC, 0);
+
+      testHierarchy(layout, 'C(Ib,T(Ia,Ic),T(Id,Ie))');
+    });
+
+    test('complex 1 b', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingItem itemC = dockingItem('c');
+      DockingItem itemD = dockingItem('d');
+      DockingItem itemE = dockingItem('e');
+      DockingColumn innerColumn = DockingColumn([itemB, itemC]);
+      DockingRow row = DockingRow([itemA, innerColumn]);
+      DockingTabs tabs = DockingTabs([itemD, itemE]);
+      DockingColumn column = DockingColumn([row, tabs]);
+      DockingLayout layout = DockingLayout(root: column);
+
+      moveItemToIndex(layout, itemA, itemC, 1);
 
       testHierarchy(layout, 'C(Ib,T(Ic,Ia),T(Id,Ie))');
     });
 
-    test('complex 2', () {
+    test('complex 2 a', () {
       DockingItem itemA = dockingItem('a');
       DockingItem itemB = dockingItem('b');
       DockingItem itemC = dockingItem('c');
@@ -125,12 +191,30 @@ void main() {
       DockingColumn column = DockingColumn([row, tabs, itemF]);
       DockingLayout layout = DockingLayout(root: column);
 
-      moveItem(layout, itemA, itemC, DropPosition.center);
+      moveItemToIndex(layout, itemA, itemC, 0);
+
+      testHierarchy(layout, 'C(Ib,T(Ia,Ic),T(Id,Ie),If)');
+    });
+
+    test('complex 2 b', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingItem itemC = dockingItem('c');
+      DockingItem itemD = dockingItem('d');
+      DockingItem itemE = dockingItem('e');
+      DockingItem itemF = dockingItem('f');
+      DockingColumn innerColumn = DockingColumn([itemB, itemC]);
+      DockingRow row = DockingRow([itemA, innerColumn]);
+      DockingTabs tabs = DockingTabs([itemD, itemE]);
+      DockingColumn column = DockingColumn([row, tabs, itemF]);
+      DockingLayout layout = DockingLayout(root: column);
+
+      moveItemToIndex(layout, itemA, itemC, 1);
 
       testHierarchy(layout, 'C(Ib,T(Ic,Ia),T(Id,Ie),If)');
     });
 
-    test('complex 3', () {
+    test('complex 3 a', () {
       DockingItem itemA = dockingItem('a');
       DockingItem itemB = dockingItem('b');
       DockingItem itemC = dockingItem('c');
@@ -145,9 +229,49 @@ void main() {
       DockingRow row2 = DockingRow([itemG, column]);
       DockingLayout layout = DockingLayout(root: row2);
 
-      moveItem(layout, itemA, itemC, DropPosition.center);
+      moveItemToIndex(layout, itemA, itemC, 0);
+
+      testHierarchy(layout, 'R(Ig,C(Ib,T(Ia,Ic),T(Id,Ie),If))');
+    });
+
+    test('complex 3 b', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingItem itemC = dockingItem('c');
+      DockingItem itemD = dockingItem('d');
+      DockingItem itemE = dockingItem('e');
+      DockingItem itemF = dockingItem('f');
+      DockingItem itemG = dockingItem('g');
+      DockingColumn innerColumn = DockingColumn([itemB, itemC]);
+      DockingRow row = DockingRow([itemA, innerColumn]);
+      DockingTabs tabs = DockingTabs([itemD, itemE]);
+      DockingColumn column = DockingColumn([row, tabs, itemF]);
+      DockingRow row2 = DockingRow([itemG, column]);
+      DockingLayout layout = DockingLayout(root: row2);
+
+      moveItemToIndex(layout, itemA, itemC, 1);
 
       testHierarchy(layout, 'R(Ig,C(Ib,T(Ic,Ia),T(Id,Ie),If))');
+    });
+
+    test('complex 3 c', () {
+      DockingItem itemA = dockingItem('a');
+      DockingItem itemB = dockingItem('b');
+      DockingItem itemC = dockingItem('c');
+      DockingItem itemD = dockingItem('d');
+      DockingItem itemE = dockingItem('e');
+      DockingItem itemF = dockingItem('f');
+      DockingItem itemG = dockingItem('g');
+      DockingColumn innerColumn = DockingColumn([itemB, itemC]);
+      DockingRow row = DockingRow([itemA, innerColumn]);
+      DockingTabs tabs = DockingTabs([itemD, itemE]);
+      DockingColumn column = DockingColumn([row, tabs, itemF]);
+      DockingRow row2 = DockingRow([itemG, column]);
+      DockingLayout layout = DockingLayout(root: row2);
+
+      moveItemToIndex(layout, itemB, tabs, 1);
+
+      testHierarchy(layout, 'R(Ig,C(R(Ia,Ic),T(Id,Ib,Ie),If))');
     });
   });
 }