+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-## 1.12.2
+## 1.12.0
+
+* `tabbed_view` dependency updated to 1.18.0
+  * Tab reordering
+* `DropPosition`
+  * The `center` value has been removed.
+* `DockingLayout`
+  * Methods `addItemOnRoot`, `addItemOn` and `moveItem`
+    * New optional parameter: `dropIndex`
+    * The `dropPostion` parameter has become optional.
+
+## 1.11.2
 
 * Bugfix
   * Error with GlobalKey when expanding a DockingTabs.